net/sched: Add API for managing shared state across multiple qdisc instances
For implementing a cross-queue shaper in mq-cake, we need a way to keep
shared state across multiple instances of sch_cake attached to different
hardware queues below the mq qdisc.

This patch introduces a generalised notion of shared qdisc state that
will be managed by the parent qdisc (in this patch implemented for mq
and mqprio), which will be assigned to the sub-qdisc when it is attached
to the parent. The shared state is keyed to a qdisc module owner (by the
qdisc->ops pointer), and will be allocated and freed by the parent.

To use this state, a qdisc module needs to set shared_size in Qdisc_ops,
and assign a function to the shared_assign method. This function will be
called whenever the qdisc is attached to a parent that manages shared
state, and will contain a pointer to an object of shared_size. The qdisc
can assign this pointer to its internal state and use it while running
for any cross-qdisc operations. When the sub-qdisc is detached, its
shared_assign method will be called with a NULL argument, at which point
it is expected to detach itself from the shared state as appropriate.

It is up to the qdisc module itself to manage concurrency across
multiple instances accessing the shared state while running; however,
calls to assign_shared are done under the RTNL, so these are not
concurrent.

A qdisc implementation can optionally assign a function pointer to the
shared_init method if it needs to do any memory initialisation of the
shared state; this method is called the first time the shared state is
allocated for a given parent instance.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>