net/sched: sch_cake: Add cake_mq qdisc for using cake on mq devices
Add a cake_mq qdisc which installs cake instances on each hardware
queue on a multi-queue device.

For now, this is just a copy of sch_mq that installs cake instead of the
default qdisc on each queue, but the idea is that the cake config will
also be shared across the sub-instances.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>